+++ leda/components/Tooltip/helpers.ts
@@ -1,16 +1,19 @@
+import { values } from 'lodash';
+
 import {
-  GetTooltipPosition, GetTooltipOffsets, TooltipPosition,
+  GetTooltipPosition, GetTooltipOffsets, TooltipPosition, CheckVerticalPosition, CheckHorizontalPosition, CheckCornerPosition,
 } from './types';
+import { Position } from './constants';
 
 /**
- * Helper checks that tooltip element is inside viewport
+ * Helper checks vertical positioning of tooltip
  * @param {DOMRect} elementRect - DOMRect object of target element
  * @param {DOMRect} tooltipRect - DOMRect object of tooltip element
  * @param {number} arrowSize - width of tooltip arrow, used as negative margin in styles of tooltip element
  *
- * @returns {boolean} - true, if tooltip element is inside viewport
+ * @returns {{isTop: boolean, isBottom: boolean, isOutsideRightBorder: boolean, isOutsideLeftBorder: boolean}}
  */
-const checkVerticalPosition = (elementRect: DOMRect, tooltipRect: DOMRect, arrowSize: number) => {
+const checkVerticalPosition: CheckVerticalPosition = (elementRect, tooltipRect, arrowSize) => {
   const isOutsideRightBorder = window.innerWidth - elementRect.right + (elementRect.width / 2) - arrowSize < (tooltipRect.width / 2);
   const isOutsideLeftBorder = elementRect.left + (elementRect.width / 2) - arrowSize < (tooltipRect.width / 2);
 
@@ -20,11 +23,22 @@ const checkVerticalPosition = (elementRect: DOMRect, tooltipRect: DOMRect, arrow
   const isBottom = (elementRect.bottom + tooltipRect.height + arrowSize <= window.innerHeight) && isInHorizontalViewPort;
 
   return {
-    isTop, isBottom, isOutsideRightBorder, isOutsideLeftBorder,
+    isTop,
+    isBottom,
+    isOutsideRightBorder,
+    isOutsideLeftBorder,
   };
 };
 
-const checkHorizontalPosition = (elementRect: DOMRect, tooltipRect: DOMRect, arrowSize: number) => {
+/**
+ * Helper checks horizontal positioning of tooltip
+ * @param {DOMRect} elementRect - DOMRect object of target element
+ * @param {DOMRect} tooltipRect - DOMRect object of tooltip element
+ * @param {number} arrowSize - width of tooltip arrow, used as negative margin in styles of tooltip element
+ *
+ * @returns {{isLeft: boolean, isRight: boolean, isOutsideTopBorder: boolean, isOutsideBottomBorder: boolean}}
+ */
+const checkHorizontalPosition: CheckHorizontalPosition = (elementRect, tooltipRect, arrowSize) => {
   const isOutsideTopBorder = elementRect.top + (elementRect.height / 2) < (tooltipRect.height / 2);
   const isOutsideBottomBorder = window.innerHeight - elementRect.bottom + (elementRect.height / 2) < (tooltipRect.height / 2);
 
@@ -34,24 +48,40 @@ const checkHorizontalPosition = (elementRect: DOMRect, tooltipRect: DOMRect, arr
   const isRight = (elementRect.right + tooltipRect.width + arrowSize <= window.innerWidth) && isInVerticalViewPort;
 
   return {
-    isLeft, isRight, isOutsideTopBorder, isOutsideBottomBorder,
+    isLeft,
+    isRight,
+    isOutsideTopBorder,
+    isOutsideBottomBorder,
   };
 };
 
-const checkCornerPosition = (isOutsideRightBorder: boolean, isOutsideLeftBorder: boolean, isOutsideTopBorder: boolean, isOutsideBottomBorder: boolean) => {
-  const isInTopLeftCorner = isOutsideTopBorder && isOutsideLeftBorder;
-  const isInTopRightCorner = isOutsideTopBorder && isOutsideRightBorder;
-  const isInBottomLeftCorner = isOutsideBottomBorder && isOutsideLeftBorder;
-  const isInBottomRightCorner = isOutsideBottomBorder && isOutsideRightBorder;
+/**
+ * Helper checks corner positioning of tooltip
+ * @param {boolean} isOutsideRightBorder - Flag defines if tooltip is outside of right border of view port
+ * @param {boolean} isOutsideLeftBorder - Flag defines if tooltip is outside of left border of view port
+ * @param {boolean} isOutsideTopBorder - Flag defines if tooltip is outside of top border of view port
+ * @param {boolean} isOutsideBottomBorder - Flag defines if tooltip is outside of bottom border of view port
+ *
+ * @returns {{isBottomCorners: boolean, isTopCorners: boolean, isRightCorners: boolean, isLeftCorners: boolean}}
+ */
+const checkCornerPosition: CheckCornerPosition = (
+  isOutsideRightBorder,
+  isOutsideLeftBorder,
+  isOutsideTopBorder,
+  isOutsideBottomBorder,
+) => {
+  const isBottomCorners = isOutsideBottomBorder && (isOutsideLeftBorder || isOutsideRightBorder);
+  const isTopCorners = isOutsideTopBorder && (isOutsideLeftBorder || isOutsideRightBorder);
+  const isRightCorners = isOutsideRightBorder && (isOutsideTopBorder || isOutsideBottomBorder);
+  const isLeftCorners = isOutsideLeftBorder && (isOutsideTopBorder || isOutsideBottomBorder);
 
   return {
-    isBottomCorners: isInBottomLeftCorner || isInBottomRightCorner,
-    isRightCorners: isInBottomRightCorner || isInTopRightCorner,
-    isLeftCorners: isInTopLeftCorner || isInBottomLeftCorner,
-    isTopCorners: isInTopLeftCorner || isInTopRightCorner,
+    isBottomCorners,
+    isTopCorners,
+    isRightCorners,
+    isLeftCorners,
   };
 };
-
 
 /**
  * Helper determines tooltip element position against target element
@@ -66,11 +96,17 @@ export const getTooltipPosition: GetTooltipPosition = ({
   position, elementRect, tooltipRect, arrowSize = 0,
 }) => {
   const {
-    isTop, isBottom, isOutsideRightBorder, isOutsideLeftBorder,
+    isTop,
+    isBottom,
+    isOutsideRightBorder,
+    isOutsideLeftBorder,
   } = checkVerticalPosition(elementRect, tooltipRect, arrowSize);
 
   const {
-    isLeft, isRight, isOutsideTopBorder, isOutsideBottomBorder,
+    isLeft,
+    isRight,
+    isOutsideTopBorder,
+    isOutsideBottomBorder,
   } = checkHorizontalPosition(elementRect, tooltipRect, arrowSize);
 
   const {
@@ -80,36 +116,36 @@ export const getTooltipPosition: GetTooltipPosition = ({
     isTopCorners,
   } = checkCornerPosition(isOutsideRightBorder, isOutsideLeftBorder, isOutsideTopBorder, isOutsideBottomBorder);
 
-  const isCornerPosition = isBottomCorners || isTopCorners;
-
   const checkPosition = (tooltipPosition: TooltipPosition): boolean => {
     switch (tooltipPosition) {
-      case 'top':
+      case Position.Top:
         return isBottomCorners || isTop;
-      case 'left':
+      case Position.Left:
         return isRightCorners || isLeft;
-      case 'right':
+      case Position.Right:
         return isLeftCorners || isRight;
-      case 'bottom':
+      case Position.Bottom:
         return isTopCorners || isBottom;
       default:
         return false;
     }
   };
 
+  const isCornerPosition = isBottomCorners || isTopCorners;
+
   if (isCornerPosition) {
-    return Array<TooltipPosition>('top', 'bottom', 'right', 'left').filter(checkPosition).join('') as TooltipPosition;
+    return values(Position).filter(checkPosition).join('-') as TooltipPosition;
   }
 
   switch (position) {
-    case 'top':
-      return Array<TooltipPosition>('top', 'bottom', 'right', 'left').find(checkPosition);
-    case 'left':
-      return Array<TooltipPosition>('left', 'right', 'bottom', 'top').find(checkPosition);
-    case 'right':
-      return Array<TooltipPosition>('right', 'left', 'top', 'bottom').find(checkPosition);
-    case 'bottom':
-      return Array<TooltipPosition>('bottom', 'top', 'left', 'right').find(checkPosition);
+    case Position.Top:
+      return Array<TooltipPosition>(Position.Top, Position.Bottom, Position.Right, Position.Left).find(checkPosition);
+    case Position.Left:
+      return Array<TooltipPosition>(Position.Left, Position.Right, Position.Bottom, Position.Top).find(checkPosition);
+    case Position.Right:
+      return Array<TooltipPosition>(Position.Right, Position.Left, Position.Top, Position.Bottom).find(checkPosition);
+    case Position.Bottom:
+      return Array<TooltipPosition>(Position.Bottom, Position.Top, Position.Left, Position.Right).find(checkPosition);
     default:
       return undefined;
   }
@@ -125,17 +161,17 @@ export const getTooltipPosition: GetTooltipPosition = ({
 export const getTooltipOffsets: GetTooltipOffsets = ({
   position, elementRect,
 }) => {
-  const includesDirection = (direction: string) => position?.includes(direction);
+  const includesPosition = (pos: string): boolean | undefined => position?.includes(pos);
 
   const top = ((): number => {
-    if (includesDirection('top')) return window.pageYOffset + elementRect.top;
-    if (includesDirection('bottom')) return window.pageYOffset + elementRect.bottom;
+    if (includesPosition(Position.Top)) return window.pageYOffset + elementRect.top;
+    if (includesPosition(Position.Bottom)) return window.pageYOffset + elementRect.bottom;
     return window.pageYOffset + elementRect.top + elementRect.height / 2;
   })();
 
   const left = ((): number => {
-    if (includesDirection('left')) return window.pageXOffset + elementRect.left;
-    if (includesDirection('right')) return window.pageXOffset + elementRect.right;
+    if (includesPosition(Position.Left)) return window.pageXOffset + elementRect.left;
+    if (includesPosition(Position.Right)) return window.pageXOffset + elementRect.right;
     return window.pageXOffset + elementRect.left + elementRect.width / 2;
   })();