+++ leda/components/Tooltip/theme.ts
@@ -1,14 +1,14 @@
 /* eslint-disable key-spacing */
 export const defaultTooltipTheme = {
   tooltip:                  'tooltip',
-  // top:                      'tooltip top',
-  // left:                     'tooltip left',
-  // right:                    'tooltip right',
-  // bottom:                   'tooltip bottom',
-  // topLeft:                  'tooltip top left',
-  // topRight:                 'tooltip top right',
-  // bottomLeft:               'tooltip bottom left',
-  // bottomRight:              'tooltip bottom right',
+  top:                      'tooltip top',
+  left:                     'tooltip left',
+  right:                    'tooltip right',
+  bottom:                   'tooltip bottom',
+  topLeft:                  'tooltip top-left',
+  topRight:                 'tooltip top-right',
+  bottomLeft:               'tooltip bottom-left',
+  bottomRight:              'tooltip bottom-right',
   wrapper:                  'tooltip-wrapper',
 };
 /* eslint-enable key-spacing */