+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - MultiSelect: Input is cleared after selecting item
+- ToolTip: Added corner position
 
 ## [0.33.0] - 2020-10-26